+++ lib/core/getters-and-setters.ts
@@ -90,7 +90,7 @@ export class GettersAndSetters<Props> implements IGettersAndSetters<Props> {
 	 *	class StringVo extends ValueObject<Props>{
 	 *		private constructor(props: Props) { super(props) }
 	 *	
-	 *		validation<Key extends keyof Props>(key: Key, value: Props[Key]): boolean {
+	 *		validation<Key extends keyof Props>(value: Props[Key], key: Key): boolean {
 	 *
 	 *			const options: IPropsValidation<Props> = {
 	 *				value: (value: string) => value.length < 15,
@@ -105,7 +105,7 @@ export class GettersAndSetters<Props> implements IGettersAndSetters<Props> {
 	 *		}
 	 *	}
 	 */
-	validation(_key: any, _value: any): boolean;
+	validation(_value: any, _key?: any): boolean;
 
 	/**
 	 * @description Validation used to `set` and `change` methods to validate value before set it.
@@ -123,7 +123,7 @@ export class GettersAndSetters<Props> implements IGettersAndSetters<Props> {
 	 *	class StringVo extends ValueObject<Props>{
 	 *		private constructor(props: Props) { super(props) }
 	 *	
-	 *		validation<Key extends keyof Props>(key: Key, value: Props[Key]): boolean {
+	 *		validation<Key extends keyof Props>(value: Props[Key], key: Key): boolean {
 	 *
 	 *			const options: IPropsValidation<Props> = {
 	 *				value: (value: string) => value.length < 15,
@@ -138,7 +138,40 @@ export class GettersAndSetters<Props> implements IGettersAndSetters<Props> {
 	 *		}
 	 *	}
 	 */
-	validation<Key extends keyof Props>(_key: Key, _value: Props[Key]): boolean { return true };
+	validation(_value: any, _key: any): boolean;
+	
+	/**
+	 * @description Validation used to `set` and `change` methods to validate value before set it.
+	 * @param _key prop key type
+	 * @param _value prop value type
+	 * @returns true if value is valid and false if is invalid.
+	 * 
+	 * 
+	 * @example
+	 * interface Props { 
+	 *		value: string;
+	 *		age: number;
+	 *	};
+	 *	
+	 *	class StringVo extends ValueObject<Props>{
+	 *		private constructor(props: Props) { super(props) }
+	 *	
+	 *		validation<Key extends keyof Props>(value: Props[Key], key: Key): boolean {
+	 *
+	 *			const options: IPropsValidation<Props> = {
+	 *				value: (value: string) => value.length < 15,
+	 *				age: (value: number) => value > 0
+	 *			} 
+	 *	
+	 *			return options[key](value);
+	 *		};
+	 *	
+	 *		public static create(props: Props): IResult<ValueObject<Props>, string> {
+	 *			return Result.Ok(new StringVo(props));
+	 *		}
+	 *	}
+	 */
+	validation<Key extends keyof Props>(_value: Props[Key], _key: Key): boolean { return true };
 
 	/**
 	 * 
@@ -181,7 +214,7 @@ export class GettersAndSetters<Props> implements IGettersAndSetters<Props> {
 					};
 				}
 
-				const canUpdate = this.validation(key, value);
+				const canUpdate = this.validation(value, key);
 				if (!canUpdate) {
 					console.log(`Trying to set value: "${value}" for key: "${String(key)}" but failed validation on ${instance?.constructor.name}`);
 					return this;
@@ -236,7 +269,7 @@ export class GettersAndSetters<Props> implements IGettersAndSetters<Props> {
 				return this
 			};
 		}
-		const canUpdate = this.validation(key, value);
+		const canUpdate = this.validation(value, key);
 		if (!canUpdate) {
 			console.log(`Trying to set value: "${value}" for key: "${String(key)}" but failed validation on ${instance?.constructor.name}`);
 			return this;